pyproject.toml
[tool.poetry]
name = "python-django-jan-2024"
version = "0.1.0"
description = ""
authors = ["SerStone <mishagoal752@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysqlclient = "^2.2.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"
daphne = "^4.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
